Reorganised modal directory.
