+++ packages/core/src/lib/modal/components/Modal.styles.ts
