+++ packages/core/src/lib/wallet-selector.types.ts
@@ -3,7 +3,7 @@ import { Observable } from "rxjs";
 import { WalletModule, Wallet } from "./wallet";
 import { WalletSelectorState } from "./store.types";
 import { Network, NetworkId } from "./options.types";
-import { ModalOptions, WalletSelectorModal } from "./modal/setupModal.types";
+import { ModalOptions, WalletSelectorModal } from "./modal/modal.types";
 import { Subscription } from "./services";
 
 export interface WalletSelectorParams {
